docker: assign non-local IP to hostname
I don't want to skip hostname check to make sure the hostname
will be used by SAP tools. I change hostname by few hacks because
"docker build" does not run in privileged mode which is necessary
for changing hostname. It is also not possible to force docker to
set hostname to build containers.